in progress

+++ mapreduce/mapreduce.go
@@ -1,7 +1,8 @@
+// Package mapreduce contains the basic types for a MapReduce process.
 package mapreduce
 
 // MapFn ...
-type MapFn func(int64, chan<- [2]string) error
+type MapFn func(int, chan<- [2]string) error
 
 // ReduceFn ...
 type ReduceFn func(string, []string) (string, error)
